move /friend routes into friendRoutes
